Added fonts/Image.NRGBA for gopherjs
Fixes #260
